Some files missing in last rename.
