+++ src/tools/qgraphicsmapscene.h
